added users and products with export and print

+++ database/migrations/2022_11_05_175501_create_promotions_table.php
@@ -20,8 +20,7 @@ class CreatePromotionsTable extends Migration
             $table->string('promotion_banner')->nullable();
             $table->string('promotion_code')->nullable();
             $table->date('promotion_expiry')->nullable();
-            $table->integer('promotion_discount')->default(0);
-
+            $table->float('promotion_discount')->default(0);
             $table->timestamps();
         });
     }
